Organize in classes and modules; A few tests added
