+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "live-py"
-version = "0.1.0-snapshot"
+version = "0.1.0"
 description = ""
 authors = ["Vitaly K. <vitaly@vitaly.com.ua>"]
 license = "LGPL-2.1-only"
@@ -11,10 +11,20 @@ mido = "^1.2.10"
 python-rtmidi = "^1.4.9"
 reactivex = "^4.0.4"
 PyYAML = "^6.0"
+coloredlogs = "^15.0.1"
 
 [tool.poetry.dev-dependencies]
 autopep8 = "^1.6.0"
 
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.2.0"
+
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.pytest.ini_options]
+log_cli = true
+log_cli_level = "DEBUG"
+log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
+log_cli_date_format = "%Y-%m-%d %H:%M:%S"
